requirements_visual.txt
Python==3.8.11
PyTorch==1.7.1
Torchvision==0.8.2
CUDA==10.2
CUDNN==7605
NumPy==1.22.3
PIL==9.2.0